AG-15239 - Update trial license form (#11274) (#11352)
* AG-15239 - Add zoom info email only form

* AG-15239 - Load zoom info if not dev

* AG-15239 - Allow trial license form to be switched out based on url search params

* AG-15239 - Add trial license all fields option

* AG-15239 - Set default trial license form in constants

* AG-15239 - Make all fields the default

* AG-15239 - Add form id to all fields trial license form

* AG-15239 - Remove useEffect to change the trial license

Didn't work with emailOnly and allFields, probably because of the ZoomInfo script. Not needed anyway.

* AG-15239 - Move zoom info to Google Tag Manager

Move Google Tag Manager configuration to env files
