+++ documentation/ag-grid-docs/.env.build
@@ -6,6 +6,11 @@ PUBLIC_ALGOLIA_SEARCH_KEY=01142fe24ea5d2e36f4eb66b0b2ff872
 PUBLIC_ALGOLIA_INDEX_PREFIX=ag-grid-dev
 PUBLIC_TRIAL_LICENCE_FORM_URL=https://us-central1-stripe-testing-19784.cloudfunctions.net/CreateLeadForTrialLK
 
+# Google Tag Manager (staging configuration)
+PUBLIC_GTM_ID=GTM-T7JG534
+PUBLIC_GTM_AUTH=K7m5Yf0NyqFbLXvp-gPogw
+PUBLIC_GTM_PREVIEW=env-55
+
 # Use production sitemap/robots disallow
 CHARTS_SITEMAP_INDEX_URL=https://www.ag-grid.com/charts/sitemap-index.xml
 CHARTS_ROBOTS_DISALLOW_JSON_URL=https://www.ag-grid.com/charts/robots-disallow.json
