+++ documentation/ag-grid-docs/.env.build.production
@@ -5,4 +5,10 @@ PUBLIC_ALGOLIA_APP_ID=O1K1ESGB5K
 PUBLIC_ALGOLIA_SEARCH_KEY=01142fe24ea5d2e36f4eb66b0b2ff872
 PUBLIC_ALGOLIA_INDEX_PREFIX=ag-grid
 PUBLIC_TRIAL_LICENCE_FORM_URL=https://us-central1-aggrid-ecommerce.cloudfunctions.net/CreateLeadForTrialLK
+
+# Google Tag Manager
+PUBLIC_GTM_ID=GTM-T7JG534
+PUBLIC_GTM_AUTH=BZRbizt3P5eacrM4bCJO5g
+PUBLIC_GTM_PREVIEW=env-1
+
 HTACCESS=true
