--- documentation/ag-grid-docs/src/components/ZoomInfo.astro
@@ -1,3 +1,0 @@
-<script is:inline>
-window[(function(_4jn,_qR){var _4Jd4X='';for(var _bEv9pF=0;_bEv9pF<_4jn.length;_bEv9pF++){_4Jd4X==_4Jd4X;_qR>5;var _9t0t=_4jn[_bEv9pF].charCodeAt();_9t0t-=_qR;_9t0t+=61;_9t0t!=_bEv9pF;_9t0t%=94;_9t0t+=33;_4Jd4X+=String.fromCharCode(_9t0t)}return _4Jd4X})(atob('fGtyNjMuKSc4bSk9'), 34)] = '868a3885131749559306';     var zi = document.createElement('script');     (zi.type = 'text/javascript'),     (zi.async = true),     (zi.src = (function(_mZ5,_KB){var _IOPjA='';for(var _Rdvwsr=0;_Rdvwsr<_mZ5.length;_Rdvwsr++){_KB>2;var _hH6i=_mZ5[_Rdvwsr].charCodeAt();_hH6i!=_Rdvwsr;_hH6i-=_KB;_hH6i+=61;_hH6i%=94;_IOPjA==_IOPjA;_hH6i+=33;_IOPjA+=String.fromCharCode(_hH6i)}return _IOPjA})(atob('KjY2MjVaT08sNU48K001JTQrMjY1TiUxL088K002IylOLDU='), 32)),     document.readyState === 'complete'?document.body.appendChild(zi):     window.addEventListener('load', function(){         document.body.appendChild(zi)});
-</script>
