+++ documentation/ag-grid-docs/src/content.config.ts
@@ -108,7 +108,6 @@ const metadata = defineCollection({
     schema: z.object({
         title: z.string(),
         description: z.string(),
-        googleTagManagerId: z.string().optional(),
         canonicalUrlBase: z.string(),
         socialImage: z.string(),
     }),
