+++ documentation/ag-grid-docs/src/content/metadata/metadata.json
@@ -1,7 +1,6 @@
 {
     "title": "AG Grid: High-Performance React Grid, Angular Grid, JavaScript Grid",
     "description": "AG Grid is a feature rich datagrid designed for the major JavaScript Frameworks. Easily integrate into your application to deliver filtering, grouping, aggregation, pivoting and much more with the performance that your users expect. Our Community version is free and open source, or take a 2 month trial of AG Grid Enterprise.",
-    "googleTagManagerId": "GTM-T7JG534",
     "canonicalUrlBase": "https://www.ag-grid.com",
     "socialImage": "/images/ag-grid-social.png"
 }
