+++ documentation/ag-grid-docs/src/layouts/Layout.astro
@@ -10,9 +10,9 @@ import Plausible from '@components/Plausible.astro';
 import { Footer } from '@ag-website-shared/components/footer/Footer';
 import styles from '@layouts/Layout.module.scss';
 import { getApiPaths } from '@utils/apiMenuPath';
-import { SITE_BASE_URL } from '@constants';
+import { SITE_BASE_URL, PUBLIC_GTM_ID, PUBLIC_GTM_AUTH, PUBLIC_GTM_PREVIEW } from '@constants';
 import GoogleTagManager from '@components/GoogleTagManager.astro';
-import ZoomInfo from '@components/ZoomInfo.astro';
+import GoogleTagManagerBody from '@components/GoogleTagManagerBody.astro';
 
 interface Props {
     title: string;
@@ -105,11 +105,18 @@ const isHomepage = Astro.url.pathname.replaceAll('/', '') === SITE_BASE_URL.repl
         <meta name="twitter:image" content={socialImage} />
 
         {
+            !isDev && (
+                <GoogleTagManager
+                    googleTagManagerId={PUBLIC_GTM_ID}
+                    googleTagManagerAuth={PUBLIC_GTM_AUTH}
+                    googleTagManagerPreview={PUBLIC_GTM_PREVIEW}
+                />
+            )
+        }
+        {
             isProduction && (
                 <Fragment>
-                    <GoogleTagManager googleTagManagerId={metadata.googleTagManagerId} />
                     <Plausible domain="ag-grid.com" />
-                    <ZoomInfo />
 
                     <script async type="text/javascript" src="https://secure.inventiveinspired7.com/js/792251.js" />
                     <noscript>
@@ -189,5 +196,15 @@ const isHomepage = Astro.url.pathname.replaceAll('/', '') === SITE_BASE_URL.repl
             )
         }
         <DevTools client:only />
+
+        {
+            !isDev && (
+                <GoogleTagManagerBody
+                    googleTagManagerId={PUBLIC_GTM_ID}
+                    googleTagManagerAuth={PUBLIC_GTM_AUTH}
+                    googleTagManagerPreview={PUBLIC_GTM_PREVIEW}
+                />
+            )
+        }
     </body>
 </html>
